+++ CHANGELOG.md
@@ -2406,3 +2406,4 @@
 - [2026-06-29 09:18] feat(forms): implement shimmer loading skeleton in forms
 - [2026-06-29 09:29] perf(core): apply RepaintBoundary on animated core widgets
 - [2026-06-29 11:07] perf(auth): apply RepaintBoundary on animated auth widgets
+- [2026-06-29 12:59] feat(analytics): add bi-directional event bus handler for analytics